requirements.txt
pandas>=2.2.2
numpy>=2.0.0
matplotlib>=3.9.0
seaborn>=0.13.2
scikit-learn>=1.5.0
boto3
python-dotenv
streamlit